+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Fixed styling on the various module settings submenu windows.
 - Deleting a connection between a Danger and an Entry on a Story Web, now works properly.  
 - You can now properly delete a Danger from a Front (right-click on the tab).
+- Fixed a number of dialogs where the enter key wouldn't activate the default button (i.e. you can now press enter to say "yes" to dialogs)
 
 ## 1.7.4 - Fixed journal entry direct-open bug
 ![](https://img.shields.io/badge/release%20date-December%2023%2C%202025-blue)